Refactor to match rest of the gift cards flow (#1308)
