+++ src/giftCards/GiftCardSettings/GiftCardExpirySettingsCard/index.ts
@@ -1,0 +1,2 @@
+export * from "./GiftCardExpirySettingsCard";
+export { default } from "./GiftCardExpirySettingsCard";
