+++ src/giftCards/GiftCardSettings/GiftCardExpirySettingsCard/messages.ts
@@ -1,6 +1,6 @@
 import { defineMessages } from "react-intl";
 
-export const giftCardSettingsListMessages = defineMessages({
+export const giftCardExpirySettingsCard = defineMessages({
   expiryDateSectionDescription: {
     defaultMessage:
       "You can set gift cards to expire after a certain time after their purchase. Remember that in some countries, gift cards expiry is prohibited by law."
