--- src/giftCards/GiftCardSettings/GiftCardSettingsList/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./GiftCardSettingsList";
-export * from "./GiftCardSettingsList";
